Linux build fix
git-svn-id: https://vast.svn.sourceforge.net/svnroot/vast@105 e31416da-b748-0410-9b4e-cac2f8189b79

+++ trunk/VAST/VAST/VASTRelay.cpp
@@ -71,7 +71,7 @@ namespace Vast
         if (curr_relay_only == false)
         {
             // set how many we send PING each time
-            int ping_num = (_relays.size () > MAX_CONCURRENT_PING ? MAX_CONCURRENT_PING : _relays.size ());
+            int ping_num = ((int)_relays.size () > MAX_CONCURRENT_PING ? MAX_CONCURRENT_PING : (int)_relays.size ());
         
             // obtain list of targets (checking if redundent PING has been sent)
             vector<bool> random_set;
